test_requirements.txt
djangocms-helper>=0.9.2,<0.10
django-filer>=1.3.0
djangocms-picture>=2.0.6
djangocms-link>=2.1.2

Pillow
html5lib>=0.999999

tox>=2.9.1
coverage>=4.4.2
flake8>=3.0.4

-e git+ssh://git@github.com/divio/djangocms-transfer.git@master#egg=djangocms-transfer
